+++ src/instance.jl
@@ -2,7 +2,7 @@
 function build(instance; root=joinpath(dirname(@__DIR__), "instances"))
     # Depot nodes
     file = joinpath(root, "$instance/depot_nodes.csv")
-    csv = CSV.File(file, types=[Int64, Int64, Float64, Float64, Int64, Float64, Float64, Float64, Float64, Float64, Float64])
+    csv = CSV.File(file, types=[Int64, Int64, Float64, Float64, Float64, Float64, Float64, Float64, Float64, Float64, Float64])
     df = DataFrame(csv)
     D = Vector{DepotNode}(undef, nrow(df))
     for k ∈ 1:nrow(df)
@@ -10,7 +10,7 @@ function build(instance; root=joinpath(dirname(@__DIR__), "instances"))
         jⁿ = df[k,2]::Int64
         x  = df[k,3]::Float64
         y  = df[k,4]::Float64
-        q  = df[k,5]::Int64
+        q  = df[k,5]::Float64
         pˡ = df[k,6]::Float64
         pᵘ = df[k,7]::Float64
         tˢ = df[k,8]::Float64
@@ -24,7 +24,7 @@ function build(instance; root=joinpath(dirname(@__DIR__), "instances"))
 
     # Customer nodes
     file = joinpath(root, "$instance/customer_nodes.csv")
-    csv = CSV.File(file, types=[Int64, Float64, Float64, Int64, Float64, Float64])
+    csv = CSV.File(file, types=[Int64, Float64, Float64, Float64, Float64, Float64])
     df = DataFrame(csv)
     ix = (df[1,1]:df[nrow(df),1])::UnitRange{Int64}
     C = OffsetVector{CustomerNode}(undef, ix)
@@ -32,7 +32,7 @@ function build(instance; root=joinpath(dirname(@__DIR__), "instances"))
         iⁿ = df[k,1]::Int64
         x  = df[k,2]::Float64
         y  = df[k,3]::Float64
-        q  = df[k,4]::Int64
+        q  = df[k,4]::Float64
         tᵉ = df[k,5]::Float64
         tˡ = df[k,6]::Float64
         iᵗ = 0
@@ -60,19 +60,19 @@ function build(instance; root=joinpath(dirname(@__DIR__), "instances"))
     end
     # Vehicles
     file = joinpath(root, "$instance/vehicles.csv")
-    csv = CSV.File(file, types=[Int64, Int64, Int64, Int64, Int64, Int64, Float64, Float64, Float64, Int64, Int64, Float64, Float64])
+    csv = CSV.File(file, types=[Int64, Int64, Int64, Float64, Float64, Float64, Float64, Float64, Float64, Float64, Int64, Float64, Float64])
     df = DataFrame(csv)
     for k ∈ 1:nrow(df)
         iᵛ = df[k,1]::Int64
         jᵛ = df[k,2]::Int64
         iᵈ = df[k,3]::Int64
-        q  = df[k,4]::Int64
-        l  = df[k,5]::Int64
-        s  = df[k,6]::Int64
+        q  = df[k,4]::Float64
+        l  = df[k,5]::Float64
+        s  = df[k,6]::Float64
         τᶠ = df[k,7]::Float64
         τᵈ = df[k,8]::Float64
         τᶜ = df[k,9]::Float64
-        τʷ = df[k,10]::Int64
+        τʷ = df[k,10]::Float64
         r̅  = df[k,11]::Int64
         πᵒ = df[k,12]::Float64
         πᶠ = df[k,13]::Float64
